requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
dnspython==2.4.2
python-whois==0.8.0
ipaddress==1.0.23
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.2

